Refactor: RentalStatus as type
